Move get_id_then_save from utils to main
Otherwise lscat app would import the api module
